+++ mt_metadata/transfer_functions/io/edi/metadata/__init__.py
@@ -5,11 +5,13 @@ from .information import Information
 from .hmeasurement import HMeasurement
 from .emeasurement import EMeasurement
 from .define_measurement import DefineMeasurement
+from .data_section import DataSection
 
 __all__ = [
     "Header",
     "Information",
     "HMeasurement",
     "EMeasurement",
-    "DefineMeasurement"
+    "DefineMeasurement",
+    "DataSection",
     ]
