+++ mt_metadata/transfer_functions/io/edi/metadata/standards/data_section.json
@@ -1,0 +1,112 @@
+{
+    "nfreq": {
+        "type": "integer",
+        "required": false,
+        "style": "number",
+        "units": null,
+        "description": "Number of frequencies",
+        "options": [],
+        "alias": [],
+        "example": "16"
+    },
+	"sectid": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "ID of the station that the data is from. This is important if you have more than one station per file.",
+        "options": [],
+        "alias": [],
+        "example": "mt001"
+    },
+	"nchan": {
+        "type": "integer",
+        "required": true,
+        "style": "number",
+        "units": null,
+        "description": "Number of channels in the transfer function",
+        "options": [],
+        "alias": [],
+        "example": "2020-01-01"
+    },
+	"maxblocks": {
+        "type": "integer",
+        "required": false,
+        "style": "number",
+        "units": null,
+        "description": "Maximum number of data blocks",
+        "options": [],
+        "alias": [],
+        "example": "cart"
+    },
+	"ex": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for EX",
+        "options": [],
+        "alias": [],
+        "example": "1"
+    },
+	"ey": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for EY",
+        "options": [],
+        "alias": [],
+        "example": "2"
+    },
+	"hx": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for HX",
+        "options": [],
+        "alias": [],
+        "example": "3"
+    },
+	"hy": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for HY",
+        "options": [],
+        "alias": [],
+        "example": "4"
+    },
+	"hz": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for HZ",
+        "options": [],
+        "alias": [],
+        "example": "5"
+    },
+	"rx": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for RRHX",
+        "options": [],
+        "alias": [],
+        "example": "6"
+    },
+	"ry": {
+        "type": "string",
+        "required": true,
+        "style": "free form",
+        "units": null,
+        "description": "Measurement ID for RRHY",
+        "options": [],
+        "alias": [],
+        "example": "7"
+    }
+}
